+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.86.7
 
+- backups: Add 'allow sealed sender from anyone' setting to AccountSettings